Do not consider bin/named/bind9.xsl.h in docs job check
The BIND 9.16 build system occasionally produces a bin/named/bind9.xsl.h
file that differs slightly from the version in the bind-9.16 branch due
to a race condition. This poses an issue for the "docs" CI job, which
executes the "maintainer-clean" make target, regenerates the
bin/named/bind9.xsl.h file, and checks for differences with the
committed file. The race condition causes discrepancies in this check.

Since this problem doesn't occur in BIND 9.18+, we can hide it by adding
bin/named/bind9.xsl.h to the list of files excluded from the check.